Finally finding bug in last_age computation
last_age was NULL for story_id not present in last run because it's value came from the joined table using FULL OUTER JOIN. I identified the bug on the first exploration of collected data, but it required me to painfully re-factor the main fetching function into smaller ones to finally find the cause of the bug.

This should be a damn reminder for me that even if it's fun to write big SQL query with CTE... in a extension context it is far better to write clean functions that can be tested separately.

So if anyone read this and is interested in writing mostly Pl/PGSQL extensions this will be my personal recommendation going forward.
- Write functions that output testable intermediates results
- It's ok to use CTE in functions, but please limit usage to small expressions or call to other testable functions
- Even if you come from SQL background acknowledge that fundamental programming principle are universal for a reason and try to write many small sub-functions instead of monolithic big SQL queries.

PS: damn still got weird discrepancy when running... guess i'll start writing tests... That's actually 2 good lessons for the price of 1 in a single commit