Improving behavior to avoid unknown and frozen stories to be harvested at once
Also introducing unknown category, this will be manually affected to story in case of an application malfunction. Instead of launching app with no data app will be able to launch with a reasonable amount of run_story with status < unknown and a large amount of unknown/frozen.

Also renaming run_ts column to ts_run for consistency