Renaming function to buil_stories_status(run_id) for clarity
This function used to be called build_stories_last which was representing only one use case. This function return the last know status of each stories for a given run_id. If that run_id is the last or is left null it happen to be the last known status.