Attempt at massive edit to swap run_id logic and use ts_run instead
Using a sequence as id has proven a very bad idea because it prevent easy merging of data.

On two occasion following a server failure it was impossible to restart the app from existing dataset.
This was to be expected in this alpha stage, and to remediate a new application was deployed and started from scratch.

Logically it should be easy to append data collected from first iteration and inject it back on new app.
On should just be concerned about correctly compute a consistent status for each story.
However the choice of sequence id for run make this trivial task quite challenging because duplicate will
exist between dataset as run_id start from value 1 as of current implementation.

To improve this behavior it has been decided to switch to a more natural and unique id  for each run.
timestamptz (ts_run) will be used as no collision should ever arise unless you try to merge dataset that cover the same timespan and it should be trivial to filter data to prevent such a case while merging.